use uint8_t instead of enum class type in RSSurfaceNodeSetSurfaceNodeType

+++ rosen/modules/render_service_client/core/ui/rs_surface_node.cpp
@@ -110,8 +110,8 @@ void RSSurfaceNode::CreateNodeInRenderThread()
 
     isChildOperationDisallowed_ = true;
     isRenderServiceNode_ = false;
-    std::unique_ptr<RSCommand> command =
-        std::make_unique<RSSurfaceNodeSetSurfaceNodeType>(GetId(), RSSurfaceNodeType::ABILITY_COMPONENT_NODE);
+    std::unique_ptr<RSCommand> command = std::make_unique<RSSurfaceNodeSetSurfaceNodeType>(GetId(),
+        static_cast<uint8_t>(RSSurfaceNodeType::ABILITY_COMPONENT_NODE));
     transactionProxy->AddCommand(command, true);
 
     // create node in RT (only when in divided render and isRenderServiceNode_ == false)
